requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-graphql-jwt==0.3.4
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==9.3.0
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tzdata==2022.7
